Use spare data from the key-generation to mutate the key part way through iterations
The aim of this is to provide some additional protection against back-tracking.

That's in _theory_ of course. In practice, I'm using a GC'd language, so there's no way to be sure the previous keys are in memory just waiting to be plucked out,

But, ignoring that, an adversary that can see current state and current output _should_ be unable to backtrack more than 24 values (24 being 1/2 an iteration loop), because the key will have been mutated based on a fragment of much, much earlier data.

The other possibility, of course, is that what we're actually doing here is keeping data in memory that could be used to re-construct values *before* the last full-key change. That would be bad