Fix a bug in the original code *and* another in the backdoor.
The bug in the original code led to key's being 0 on mutate - because "spare" contains the full block, the first 32 bytes are the same as the key. An XOR is therefore going to return 0

Once I fixed that, continued to get 0'd keys back but at different intervals. When deriving a key with the backdoor in place we need to skip the injected bytes, otherwise we end up eventually XORing our original key against itself

Attack still works but it's no longer a 0'd key

    pi@nextcloudpi:~ $ wc -l output
    128 output
    pi@nextcloudpi:~ $ python3 attack.py
    Got 128 blocks
    2752 greater than 2688. Abort
    Key is 32 bytes
    Found key b'tTbFGHGbbW09xbIAQXdpjWRaHv1gkaFqpEm888r5h4s='
    Successfully predicted that position -6 would contain b'aqN7KwG3WFeFhqrXEnGpjxKwc0nNI7Zh+/H/asWWq1bDOJzA+1U8Nf62P3sZi5qf9nCNHaSXzTtDyqfXCcJl/g==' (b'aqN7KwG3WFeFhqrXEnGpjxKwc0nNI7Zh+/H/asWWq1bDOJzA+1U8Nf62P3sZi5qf9nCNHaSXzTtDyqfXCcJl/g==')