Adjust backdoor so that previous key is used as the seed value for the next round.
This discretely communicates the key across the re-keying threshold but isn't enough for our back-track to jump the threshold yet.

We'd need the final bytes output from that block too, so that we can decrypt those with the key.

The problem here now, is one of trying to communicate 128 bytes of hard-to-compress information with only 64 bytes in which to do it.